requirements.txt
requests==2.31.0
lxml==4.9.3
tqdm==4.65.0
Pillow==9.5.0